requirements.txt
beanie~=1.23.6
aiohttp~=3.8.6
discord-py-interactions
Pillow~=10.1.0
python-dotenv~=1.0.0
pymongo~=4.6.1
pyyaml~=6.0.1
motor~=3.3.2